refactor: JsonHighlight を shared/ui に分離
